+++ src/com/thecalculator/Subtract.java
@@ -2,5 +2,5 @@ package com.thecalculator;
 
 public interface Subtract {
 
-    int subtract(int number);
+    long subtract(long number);
 }
